requirements.txt
beautifulsoup4==4.13.4
chroma-hnswlib==0.7.3
chromadb==1.0.12
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
python-dotenv==1.1.0
selenium==4.33.0
streamlit==1.45.1